feat: UserDetails, constants.ts, users.service.ts
Users can change their personal information, inlcuding their password. Validations are in place. If a fild is invalid, user will be notified under that field. Button is disabled unless there is change in the form. On submit, the submit button will become a 'loading' button. When user data is successfully updated a 'success' alert will be shown over the 'update info' and 'cancel buttons'. Users will not be able to upload any file as an image, onlt jpeg, png. NOTE: Because Apple are special and some browser don't support those type of image files, the check for a HEIC file is if the file type is ''. Could cause problems...