feat: remove chat
Base functionality for removing a chat for yourself. If it is a more than two members chat you remove yourself from the chat and other members see that. If it is a chat between two members the chat is removed from your chats till you make a chat with this user again or he writes to you back.